requirements.txt
requests>=2.25.0
PyNaCl>=1.5.0
cryptography>=3.4.0
azure-identity 
azure-mgmt-authorization 
azure-mgmt-resource 
msgraph-sdk